requirements.txt
flask_restplus==0.13.0
Flask==1.1.2
numpy==1.19.5
Flask_Cors==3.0.10
opencv_python==4.5.1.48
onnxruntime==1.7.0
Werkzeug==0.16.1
Pillow==8.1.2
gunicorn==20.0.4
